Compat EF - Remove `ace_hearing` hard dependency (#10646)
