+++ addons/compat_ef/CfgWeapons.hpp
@@ -7,25 +7,6 @@ class CfgWeapons {
         EGVAR(nightvision,border) = QPATHTOEF(nightvision,data\nvg_mask_binos_4096.paa);
     };
 
-    // -- ace_hearing --
-    class HelmetBase;
-    class EF_H_Protecta: HelmetBase {
-        HEARING_PROTECTION_PELTOR;
-    };
-    class EF_H_MCH; // this does not have peltor
-    class EF_H_MCH_Basic: EF_H_MCH {
-        HEARING_PROTECTION_PELTOR;
-    };
-    class EF_H_MCH_BasicNet_Des: EF_H_MCH {
-        HEARING_PROTECTION_PELTOR;
-    };
-    class EF_H_MCH_Full: EF_H_MCH {
-        HEARING_PROTECTION_PELTOR;
-    };
-    class EF_H_MCH_FullCamo_Des: EF_H_MCH {
-        HEARING_PROTECTION_PELTOR;
-    };
-
     // -- ace_overpressure --
     class autocannon_30mm;
     class EF_autocannon_50mm_AAV9: autocannon_30mm {
