+++ addons/compat_ef/compat_ef_hearing/CfgWeapons.hpp
@@ -1,0 +1,19 @@
+class CfgWeapons {
+    class HelmetBase;
+    class EF_H_Protecta: HelmetBase {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class EF_H_MCH; // this does not have peltor
+    class EF_H_MCH_Basic: EF_H_MCH {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class EF_H_MCH_BasicNet_Des: EF_H_MCH {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class EF_H_MCH_Full: EF_H_MCH {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class EF_H_MCH_FullCamo_Des: EF_H_MCH {
+        HEARING_PROTECTION_PELTOR;
+    };
+};
