+++ addons/compat_ef/compat_ef_hearing/config.cpp
@@ -1,16 +1,18 @@
 #include "script_component.hpp"
+#include "\z\ace\addons\hearing\script_macros_hearingProtection.hpp"
 
 class CfgPatches {
-    class ADDON {
+    class SUBADDON {
         name = COMPONENT_NAME;
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ace_common", "EF_Weapons", "EF_Marines"};
+        requiredAddons[] = {"EF_Weapons", "EF_Marines", "ace_hearing"};
         skipWhenMissingDependencies = 1;
         author = ECSTRING(common,ACETeam);
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
+        addonRootClass = QUOTE(ADDON);
     };
 };
